testing the provider with the mockitojunitrunner
