+++ todoapp/app/src/test/java/com/example/android/architecture/blueprints/todoapp/data/source/MockCursorProvider.java
@@ -77,6 +77,8 @@ public class MockCursorProvider {
         return new TaskMockCursor(entryList);
     }
 
+
+
     public static class TaskMockCursor extends MockCursor {
         Map<Integer, Object> entry;
         int cursorIndex;
