+++ todoapp/app/src/test/java/com/example/android/architecture/blueprints/todoapp/data/source/MockTasksProvider.java
@@ -1,0 +1,9 @@
+package com.example.android.architecture.blueprints.todoapp.data.source;
+
+import android.test.mock.MockContentProvider;
+
+public class MockTasksProvider extends MockContentProvider {
+
+
+
+}
